+++ src/apps/hedgefarm/index.ts
@@ -1,0 +1,3 @@
+export { HEDGEFARM_DEFINITION, HedgefarmAppDefinition } from './hedgefarm.definition';
+export { HedgefarmAppModule } from './hedgefarm.module';
+export { HedgefarmContractFactory } from './contracts';