+++ src/apps/homora-v2/interfaces/enums.ts
@@ -1,0 +1,50 @@
+export enum Exchange {
+  UniswapV2 = 'Uniswap V2',
+  UniswapV3 = 'Uniswap V3',
+  Sushiswap = 'Sushiswap',
+  Curve = 'Curve',
+  Balancer = 'Balancer',
+  TraderJoe = 'Trader Joe',
+  Pangolin = 'Pangolin',
+  PangolinV2 = 'Pangolin V2',
+  SpookySwap = 'SpookySwap',
+  SpookySwapV2 = 'SpookySwap V2',
+  SpiritSwap = 'SpiritSwap',
+  Beethovenx = 'Beethovenx',
+}
+
+export enum PoolType {
+  LiquidityProviding = 'Liquidity Providing',
+  YieldFarming = 'Yield Farming',
+}
+
+export enum Poolstatus {
+  delisted = 'delisted',
+  delisting = 'delisting',
+  emergency_delisted = 'emergency_delisted',
+}
+
+export enum WTokenType {
+  WERC20 = 'WERC20',
+  WStakingRewards = 'WStakingRewards',
+  WMasterChef = 'WMasterChef',
+  WMasterChefJoeV3 = 'WMasterChefJoeV3',
+  WBoostedMasterChefJoe = 'WBoostedMasterChefJoe',
+  WLiquidityGauge = 'WLiquidityGauge',
+  WMiniChef = 'WMiniChef',
+  WMiniChefV2 = 'WMiniChefV2',
+  WMasterChefBeets = 'WMasterChefBeets',
+  WMasterChefBoolV2 = 'WMasterChefBoolV2',
+  WUniswapV3 = 'WUniswapV3',
+}
+
+export enum UniswapV3Range {
+  Wide = 'Wide',
+  Medium = 'Medium',
+  Narrow = 'Narrow',
+}
+
+export enum SafeboxStatus {
+  delisted = 'delisted',
+  delisting = 'delisting',
+}